refactor: productStore 경로 수정 및 관련 파일 업데이트
