+++ src/features/product/model/product-store.ts
